Renamed ytm.py & added README to tests
> Why Additions/Deletions/Changes?

ytm.py > youtube_music.py: 'ytm' is not a very descriptive name.

README.md: Running tests with mocked recordings isn't exactly very
obvious. Considering that there are 2 different plugins that allow for
mocked network requests (pytest-recording and pytest-vcr) it felt
correct to include a short section on handling the mocked requests.

the rest: adjustments to handle the ytm.py rename

> How it helps?

ytm.py: the name is more descriptive now and will also be more in-line
with upcoming modules (spotdl.defaults.download.youtube_music) that is
being worked on as of writing this commit message

README.md: outlines the most common usecases of pytest-recording. You
don't have to run a few google searches to work with it anymore

the rest: necessary changes to keep the tests working after renaming
ytm.py
